Refactorización del componente FlashCardGenerator:, añadiendo hook useForm
