+++ src/components/generator/GeneratorFlashCard.tsx
@@ -1,53 +1,19 @@
-import { useRef, useState } from "react";
-import { validatePregunta } from "@/utils/schemes/formValidation";
 import { useUser } from "@clerk/nextjs";
 import { useFlashcardSync } from "@/hooks/useFlashcardSync";
 import SyncIndicator  from "../ui/sync-indicator";
-import debounce from "debounce";
 import { useGetAnswers } from "@/hooks/useGetAnswers";
 import { usePing } from "@/hooks/usePing";
 import LoadingModal from "../fallbacks/LoadingModal";
 import { FlashCardGenerator } from "./FlashCardGenerator";
 
 export const Generator = () => {
-  const [pregunta, setPregunta] = useState("");
-  const [error, setError] = useState<null | string>(null);
-  const textareaRef = useRef<HTMLTextAreaElement>(null);
   const { user, isLoaded } = useUser();
-
-
   const user_id = user?.id;
   const { loadingAnswers } = useGetAnswers();
 
   usePing();
-
- 
-
   // Iniciar sincronización
   useFlashcardSync(user_id as string);
-
-  const handlePreguntaChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    const value = e.target.value;
-    setPregunta(value);
-
-    const errorMessage = validatePregunta(value);
-    debouncedSetError(errorMessage);
-
-   /* if (textareaRef.current) {
-      textareaRef.current.style.height = "auto"; // Reinicia el alto
-      textareaRef.current.style.height =
-        textareaRef.current.scrollHeight + "px"; // Ajusta al contenido
-    } */
-  };
-
-  /*const resetForm = () => {
-    setPregunta("");
-    if (textareaRef.current) {
-      textareaRef.current.style.height = "auto";
-    }
-  };*/
-
- 
 
   // Mostrar mensaje de carga si el usuario aún no está listo
   if (!isLoaded || !user || !user.id) {
@@ -79,11 +45,7 @@ export const Generator = () => {
       ) : (
         <FlashCardGenerator
           user={user}
-          error={error}
-          pregunta={pregunta}
-          textAreaRef={textareaRef}
           loadingAnswers={loadingAnswers}
-          handlePreguntaChange={handlePreguntaChange}
         />
         )}
       <SyncIndicator />
